fix: Add watchdog to detect stale Redis connections in activation workers
Add watchdog thread to ActivationWorker that monitors Redis queue
registration and forces process exit if the activation worker becomes
unhealthy due to stale TCP connections after a Redis crash or
intermittent network connectivity that disconnects the activation
workers.

The watchdog checks every (DEFAULT_WORKER_HEARTBEAT_TIMEOUT / 2)
seconds if the worker's queues are registered in Redis. If
registration is incomplete for 1.5x the heartbeat timeout, it calls
os._exit(1) to trigger orchestrator restart. We check the queues
twice per heartbeat, while giving it a 50% grace period to recover.

Essentially, we give the process 3 attempts to recover, considering
that:
  - 1 failure: could be a transient network blip
  - 2 failures: still possibly recovering
  - 3 failures: definite problem; exit and restart

Only enabled for standalone Redis deployments (not HA clusters), and
also only applicable to EDA running in OpenShift/Kubernetes, or
container-based; e.g.: Redis is remote, and not running locally or not
behind a Unix socket. In these cases, there are no stale TCP
connections.